feat: add evaluation results page
