+++ src/features/evaluation-results/[id]/view-evaluation-results-header.tsx
@@ -1,57 +1,190 @@
-import { useEffect } from "react"
-import { useAppSelector } from "../../../hooks/useAppSelector"
-import { useParams } from "react-router-dom"
-import { formatDateRange } from "../../../utils/format-date"
-import { useAppDispatch } from "../../../hooks/useAppDispatch"
-import { PageTitle } from "../../../components/shared/page-title"
-import { getCmEvaluationResult } from "../../../redux/slices/evaluation-result-slice"
-import { ScoreRange } from "../../../components/shared/score-range/score-range"
+import { useEffect, useState } from "react"
+import { useSearchParams } from "react-router-dom"
+import { Input } from "../../components/ui/input/input"
+import { CustomSelect } from "../../components/ui/select/custom-select"
+import { Button } from "../../components/ui/button/button"
+import { type Option } from "../../types/optionType"
+import { useAppDispatch } from "../../hooks/useAppDispatch"
+import { getEvaluationAdministrations } from "../../redux/slices/evaluation-administrations-slice"
+import { useAppSelector } from "../../hooks/useAppSelector"
+import { EvaluationAdministrationStatus } from "../../types/evaluation-administration-type"
+import { getScoreRatings } from "../../redux/slices/score-ratings-slice"
+import { Banding } from "../../types/banding-type"
 
-export const ViewEvaluationResultsHeader = () => {
-  const { id } = useParams()
+const bandingFilters: Option[] = Object.values(Banding).map((value) => ({
+  label: value,
+  value,
+}))
+
+bandingFilters.unshift({
+  label: "All",
+  value: "all",
+})
+
+const sortByFilters: Option[] = [
+  {
+    label: "All",
+    value: "all",
+  },
+  {
+    label: "Evaluee",
+    value: "evaluee",
+  },
+  {
+    label: "Score [Low to High]",
+    value: "score_asc",
+  },
+  {
+    label: "Score [High to Low]",
+    value: "score_desc",
+  },
+  {
+    label: "Standard Score [Low to High]",
+    value: "standard_score_asc",
+  },
+  {
+    label: "Standard Score [High to Low]",
+    value: "standard_score_desc",
+  },
+]
+
+export const EvaluationResultsListFilter = () => {
+  const [searchParams, setSearchParams] = useSearchParams()
+
   const appDispatch = useAppDispatch()
-  const { evaluation_result } = useAppSelector((state) => state.evaluationResult)
+  const { evaluation_administrations } = useAppSelector((state) => state.evaluationAdministrations)
+  const { score_ratings } = useAppSelector((state) => state.scoreRatings)
+
+  const [evaluationAdministrationFilters, setEvaluationAdministrationFilters] = useState<Option[]>(
+    []
+  )
+  const [scoreRatingFilters, setScoreRatingFilters] = useState<Option[]>([])
+
+  const [name, setName] = useState<string>(searchParams.get("name") ?? "")
+  const [evaluationAdministrationId, setEvaluationAdministrationId] = useState<string>(
+    searchParams.get("evaluation_administration_id") ?? "all"
+  )
+  const [scoreRatingId, setScoreRatingId] = useState<string>(
+    searchParams.get("score_ratings_id") ?? "all"
+  )
+  const [banding, setBanding] = useState<string>(searchParams.get("banding") ?? "all")
+  const [sortBy, setSortBy] = useState<string>(searchParams.get("sort_by") ?? "all")
 
   useEffect(() => {
-    if (id !== undefined) {
-      void appDispatch(getCmEvaluationResult(parseInt(id)))
-    }
+    void appDispatch(
+      getEvaluationAdministrations({
+        status: [
+          EvaluationAdministrationStatus.Closed,
+          EvaluationAdministrationStatus.Published,
+        ].join(","),
+      })
+    )
+    void appDispatch(getScoreRatings())
   }, [])
 
+  useEffect(() => {
+    const filterOptions: Option[] = evaluation_administrations.map((evalAdmin) => ({
+      label: evalAdmin.name ?? "",
+      value: evalAdmin.id.toString(),
+    }))
+    filterOptions.unshift({
+      label: "All",
+      value: "all",
+    })
+    setEvaluationAdministrationFilters(filterOptions)
+  }, [evaluation_administrations])
+
+  useEffect(() => {
+    const filterOptions: Option[] = score_ratings.map((scoreRating) => ({
+      label: scoreRating.name ?? "",
+      value: scoreRating.id.toString(),
+    }))
+    filterOptions.unshift({
+      label: "All",
+      value: "all",
+    })
+    setScoreRatingFilters(filterOptions)
+  }, [score_ratings])
+
+  const handleSearch = async () => {
+    if (name.length !== 0) {
+      searchParams.set("name", name)
+    } else {
+      searchParams.delete("name")
+    }
+    searchParams.set("evaluation_administration_id", evaluationAdministrationId)
+    searchParams.set("score_ratings_id", scoreRatingId)
+    searchParams.set("banding", banding)
+    searchParams.set("sort_by", sortBy)
+    searchParams.set("page", "1")
+    setSearchParams(searchParams)
+  }
+
+  const handleClear = async () => {
+    setName("")
+    setEvaluationAdministrationId("all")
+    setScoreRatingId("all")
+    setBanding("all")
+    setSortBy("all")
+    setSearchParams({})
+  }
+
   return (
-    <>
-      <div className='flex flex-col'>
-        <div className='flex flex-col justify-between items-start md:items-end mt-2 md:flex-row gap-5'>
-          <div>
-            <div className='flex gap-4 primary-outline items-end mb-4'>
-              <PageTitle>Evaluation Results</PageTitle>
-            </div>
-            <div className='flex gap-3 font-bold'>
-              {evaluation_result?.users?.last_name}, {evaluation_result?.users?.first_name}
-            </div>
-            <div className='flex gap-3'>{evaluation_result?.eval_admin_name}</div>
-            <div className='flex gap-3'>
-              {formatDateRange(
-                evaluation_result?.eval_period_start_date,
-                evaluation_result?.eval_period_end_date
-              )}
-            </div>
-          </div>
+    <div className='flex flex-col md:flex-row justify-between gap-4'>
+      <div className='flex-1 flex flex-col md:flex-row gap-4'>
+        <div className='flex-1'>
+          <Input
+            label='Name'
+            name='search'
+            placeholder='Search by name'
+            value={name}
+            onChange={(e) => setName(e.target.value)}
+          />
         </div>
-        <div className='text-xl text-primary-500 font-bold my-5'>
-          Total Score: {evaluation_result?.total_score}%
-        </div>
-        {evaluation_result?.score_rating !== undefined &&
-          evaluation_result?.users?.picture !== undefined && (
-            <ScoreRange
-              user_picture={evaluation_result?.users?.picture}
-              score_rating={evaluation_result?.score_rating}
-              score={evaluation_result?.score}
-              size='medium'
-              is_evaluee={false}
-            />
+        <CustomSelect
+          data-test-id='EvaluationAdministration'
+          label='Evaluation Administration'
+          name='evaluation_administration'
+          value={evaluationAdministrationFilters.find(
+            (option) => option.value === evaluationAdministrationId
           )}
+          onChange={(option) =>
+            setEvaluationAdministrationId(option !== null ? option.value : "all")
+          }
+          options={evaluationAdministrationFilters}
+          fullWidth
+        />
+        <CustomSelect
+          data-test-id='ScoreRating'
+          label='Rating'
+          name='score_rating'
+          value={scoreRatingFilters.find((option) => option.value === scoreRatingId)}
+          onChange={(option) => setScoreRatingId(option !== null ? option.value : "all")}
+          options={scoreRatingFilters}
+        />
+        <CustomSelect
+          data-test-id='Banding'
+          label='Banding'
+          name='banding'
+          value={bandingFilters.find((option) => option.value === banding)}
+          onChange={(option) => setBanding(option !== null ? option.value : "all")}
+          options={bandingFilters}
+        />
+        <CustomSelect
+          data-test-id='SortBy'
+          label='Sort by'
+          name='sort_by'
+          value={sortByFilters.find((option) => option.value === sortBy)}
+          onChange={(option) => setSortBy(option !== null ? option.value : "all")}
+          options={sortByFilters}
+        />
       </div>
-    </>
+      <div className='flex items-end gap-4'>
+        <Button onClick={handleSearch}>Search</Button>
+        <Button variant='primaryOutline' onClick={handleClear}>
+          Clear
+        </Button>
+      </div>
+    </div>
   )
 }
