+++ src/features/evaluation-results/[id]/view-evaluation-results-table.tsx
@@ -1,116 +1,84 @@
-import { useState } from "react"
-import { useParams } from "react-router-dom"
-import { useAppSelector } from "../../../hooks/useAppSelector"
-import { Loading } from "../../../types/loadingType"
-import Dialog from "../../../components/ui/dialog/dialog"
-import { Button } from "../../../components/ui/button/button"
-import { type EvaluationResultDetail } from "../../../types/evaluation-result-detail-type"
-import { Progress } from "../../../components/ui/progress/progress"
-import { getScoreVariant } from "../../../utils/variant"
+import { useEffect } from "react"
+import { useNavigate, useSearchParams } from "react-router-dom"
+import { useAppDispatch } from "../../hooks/useAppDispatch"
+import { useAppSelector } from "../../hooks/useAppSelector"
+import { getCmEvaluationResults } from "../../redux/slices/evaluation-results-slice"
+import { Pagination } from "../../components/shared/pagination/pagination"
+import { formatDate } from "../../utils/format-date"
 
-export const ViewEvaluationResultsTable = () => {
-  const { id } = useParams()
-  const { loading, evaluation_result } = useAppSelector((state) => state.evaluationResult)
-  const [showDetails, setShowDetails] = useState<boolean>(false)
-  const [selectedEvaluationResultDetail, setSelectedEvaluationResultDetail] =
-    useState<EvaluationResultDetail>()
+export const EvaluationResultsListTable = () => {
+  const navigate = useNavigate()
+  const [searchParams] = useSearchParams()
 
-  const toggleDetails = () => {
-    setShowDetails((prev) => !prev)
-  }
+  const appDispatch = useAppDispatch()
 
-  const handleOpenDetails = (detail: EvaluationResultDetail) => {
-    setSelectedEvaluationResultDetail(detail)
+  const { evaluation_results, hasPreviousPage, hasNextPage, totalPages } = useAppSelector(
+    (state) => state.evaluationResults
+  )
+
+  useEffect(() => {
+    void appDispatch(
+      getCmEvaluationResults({
+        name: searchParams.get("name") ?? undefined,
+        evaluation_administration_id: searchParams.get("evaluation_administration_id") ?? undefined,
+        score_ratings_id: searchParams.get("score_ratings_id") ?? undefined,
+        banding: searchParams.get("banding") ?? undefined,
+        sort_by: searchParams.get("sort_by") ?? undefined,
+        page: searchParams.get("page") ?? undefined,
+      })
+    )
+  }, [searchParams])
+
+  const handleViewEvaluationResult = (id: number) => {
+    navigate(`/evaluation-results/${id}`)
   }
 
   return (
-    <>
-      {loading === Loading.Pending && <div>Loading...</div>}
-      {loading === Loading.Fulfilled && evaluation_result === null && <div>Not found</div>}
-      <div>
-        {loading === Loading.Fulfilled &&
-          evaluation_result?.evaluation_result_details !== undefined &&
-          evaluation_result?.evaluation_result_details.length > 0 &&
-          id !== undefined && (
-            <>
-              <div className='text-xl text-primary-500 font-bold mb-5'>Detailed Evaluation </div>
-              <table className='md:w-[860px] table-fixed'>
-                <thead className='text-left'>
-                  <tr>
-                    <th className='py-1 border-b-4 text-primary-500'>Evaluations</th>
-                    <th className='py-1 border-b-4 text-start text-primary-500'>Score</th>
-                    <th className='py-1 border-b-4 text-start text-primary-500'>Rating</th>
-                    <th className='py-1 border-b-4 text-center text-primary-500'>Standard Score</th>
-                    <th className='py-1 border-b-4 text-start text-primary-500'>Banding</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {evaluation_result?.evaluation_result_details.map((detail) => (
-                    <tr
-                      key={detail.id}
-                      className='cursor-pointer hover:bg-slate-100'
-                      onClick={() => {
-                        handleOpenDetails(detail)
-                        toggleDetails()
-                      }}
-                    >
-                      <td className='py-1 border-b'>{detail.template_name}</td>
-                      <td className='py-1 border-b text-start'>{detail.total_score}%</td>
-                      {detail.score_rating?.display_name !== null && (
-                        <td className='py-1 border-b text-start items-center'>
-                          {detail.score_rating?.display_name}
-                        </td>
-                      )}
-                      <td className='py-1 border-b text-center items-center'>
-                        {Number(detail.zscore).toFixed(2)}
-                      </td>
-                      <td className='py-1 border-b text-start items-center'>{detail.banding}</td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </>
-          )}
-      </div>
-      <Dialog open={showDetails} size={"medium"} maxWidthMin={true}>
-        <Dialog.Title>
-          <div className='py-1 text-primary-500'>
-            {selectedEvaluationResultDetail?.template_name} Score:{" "}
-            {selectedEvaluationResultDetail?.total_score}%
-          </div>
-        </Dialog.Title>
-        <Dialog.Description>
-          <div className='overflow-auto'>
-            {selectedEvaluationResultDetail?.evaluation_template_contents?.map((content, index) => (
-              <div key={index} className='hover:bg-slate-100 p-2'>
-                <div className='flex justify-between mb-2'>
-                  <div className='text-primary-500 font-bold w-1/2'>{content.name}</div>
-                  <div className='w-[300px] relative'>
-                    <div className='relative z-0'>
-                      <Progress
-                        variant={getScoreVariant(content.average_rate ?? 0)}
-                        value={content.average_rate ?? 0}
-                        width='w-[20px]'
-                      />
-                    </div>
-                    <div
-                      className={`absolute top-[10px] right-3/4 transform -translate-x-1/2 -translate-y-1/2 z-10  text-white`}
-                    >
-                      {content.average_rate}%
-                    </div>
-                  </div>
-                </div>
-                <div className='mb-2 text-sm italic'> {content.description}</div>
-              </div>
-            ))}
-          </div>
-        </Dialog.Description>
-        <Dialog.Actions>
-          <Button variant='primary' onClick={toggleDetails}>
-            Close
-          </Button>
-        </Dialog.Actions>
-      </Dialog>
-    </>
+    <div className='flex flex-col gap-8'>
+      <table className='w-full table-fixed'>
+        <thead className='text-left'>
+          <tr>
+            <th className='pb-3'>Evaluee Name</th>
+            <th className='pb-3'>Eval Admin Name</th>
+            <th className='pb-3'>Eval Period</th>
+            <th className='pb-3'>Score</th>
+            <th className='pb-3'>Score Rating</th>
+            <th className='pb-3'>Standard Score</th>
+            <th className='pb-3'>Banding</th>
+          </tr>
+        </thead>
+        <tbody>
+          {evaluation_results.map((evaluationResult) => (
+            <tr
+              className='cursor-pointer hover:bg-slate-100'
+              key={evaluationResult.id}
+              onClick={() => handleViewEvaluationResult(evaluationResult.id)}
+            >
+              <td className='py-1'>
+                {evaluationResult.users?.last_name}, {evaluationResult.users?.first_name}
+              </td>
+              <td className='py-1'>{evaluationResult.evaluation_administration?.name}</td>
+              <td className='py-1'>
+                {formatDate(evaluationResult.evaluation_administration?.eval_period_start_date)} to{" "}
+                {formatDate(evaluationResult.evaluation_administration?.eval_period_end_date)}
+              </td>
+              <td className='py-1'>{evaluationResult.score}</td>
+              <td className='py-1'>{evaluationResult.score_ratings?.display_name}</td>
+              <td className='py-1'>{evaluationResult.zscore}</td>
+              <td className='py-1'>{evaluationResult.banding}</td>
+            </tr>
+          ))}
+        </tbody>
+      </table>
+      {totalPages !== 1 && (
+        <div className='flex justify-center'>
+          <Pagination
+            hasPreviousPage={hasPreviousPage}
+            hasNextPage={hasNextPage}
+            totalPages={totalPages}
+          />
+        </div>
+      )}
+    </div>
   )
 }
