--- src/features/evaluation-results/evaluation-results-header.tsx
@@ -1,9 +1,0 @@
-import { PageTitle } from "../../components/shared/page-title"
-
-export const EvaluationResultsHeader = () => {
-  return (
-    <div className='flex flex-col md:flex-row justify-between gap-4'>
-      <PageTitle>Evaluation Results</PageTitle>
-    </div>
-  )
-}
