+++ src/pages/evaluation-results/[id]/view-evaluation-results.tsx
@@ -1,9 +1,7 @@
 import { ViewEvaluationResultsHeader } from "../../../features/evaluation-results/[id]/view-evaluation-results-header"
 import { ViewEvaluationResultsTable } from "../../../features/evaluation-results/[id]/view-evaluation-results-table"
-import { ViewEvaluationResultsChart } from "../../../features/evaluation-results/[id]/view-evaluation-results-chart"
 import { ViewEvaluationResultsComments } from "../../../features/evaluation-results/[id]/view-evaluation-results-comments"
 import { ViewEvaluationResultsFooter } from "../../../features/evaluation-results/[id]/view-evaluation-results-footer"
-import { ViewEvaluationResultsAttendanceAndPunctuality } from "../../../features/evaluation-results/[id]/view-evaluation-results-attendance-and-punctuality"
 
 export default function ViewEvaluationResults() {
   return (
@@ -11,8 +9,6 @@ export default function ViewEvaluationResults() {
       <div className='h-[calc(100vh_-_104px)] flex flex-col gap-8'>
         <ViewEvaluationResultsHeader />
         <ViewEvaluationResultsTable />
-        <ViewEvaluationResultsAttendanceAndPunctuality />
-        <ViewEvaluationResultsChart />
         <ViewEvaluationResultsComments />
         <ViewEvaluationResultsFooter />
       </div>
