+++ src/routes/cm-user-route.tsx
@@ -2,8 +2,12 @@ import { lazy } from "react"
 import { Navigate, Outlet } from "react-router-dom"
 import { useCmUser } from "../hooks/use-cm-user"
 
-const EvaluationResults = lazy(
-  async () => await import("../pages/evaluation-results/evaluation-results")
+const EvaluationResultsList = lazy(
+  async () => await import("../pages/evaluation-results/evaluation-results-list")
+)
+
+const ViewEvaluationResults = lazy(
+  async () => await import("../pages/evaluation-results/[id]/view-evaluation-results")
 )
 
 export const cmUserRoutes = {
@@ -14,7 +18,11 @@ export const cmUserRoutes = {
      */
     {
       path: "/evaluation-results",
-      element: <EvaluationResults />,
+      element: <EvaluationResultsList />,
+    },
+    {
+      path: "/evaluation-results/:id",
+      element: <ViewEvaluationResults />,
     },
   ],
 }
